Updated to use different MinZ
This time it does an iterated select over the table, grabbing the first item that is smaller than the current one, and then going again until there is nothing smaller.

It also does a select_delete on any larger items, so that the search space has been shrunk next time around.  I'm not sure if that actually makes things more efficient, but I hope it does.