Fix error in finding min
Turned out I had the match specification slightly incorrect, with the result that it was always returning '$end_of_table' after the very first result

Also, experimented with the select_delete in the findMinZ function.  Looks like with a fully-connected graph with 10 nodes, it leads to a ~100ms speed improvement.